+++ actionview/test/lib/test_renderable.rb
@@ -1,13 +1,11 @@
 # frozen_string_literal: true
 
 class TestRenderable
-  def render_in(view_context, **)
+  def render_in(view_context, locals: {}, **)
     if block_given?
       view_context.render(html: yield)
     else
-      view_context.render(inline: <<~ERB.strip, **)
-        Hello, <%= local_assigns[:name] || "World" %>!
-      ERB
+      "Hello, #{locals[:name] || "World"}!"
     end
   end
 end